move join function resolving to module
