+++ crates/polars-plan/src/plans/conversion/mod.rs
@@ -22,6 +22,7 @@ use polars_utils::vec::ConvertVec;
 use recursive::recursive;
 mod functions;
 pub(crate) mod type_coercion;
+mod join;
 
 pub(crate) use expr_expansion::{expand_selectors, is_regex_projection, prepare_projection};
